requirements.txt
opencv-python
Pillow
torchmetrics
easydict
six
scipy
matplotlib
